% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_binary_column}
\alias{add_binary_column}
\title{Add Binary Column to Data Frame}
\usage{
add_binary_column(df_name, column_name, value)
}
\arguments{
\item{df_name}{The name of the data frame as a character string.}

\item{column_name}{The name of the column to check.}

\item{value}{The value to check for.}
}
\value{
No return value. Modifies the data frame in the global environment.
}
\description{
This function adds a binary column to a data frame based on the value of another column.
}
\examples{
df <- data.frame(observation = c("bird", "dog", "cat", "dog"))
add_binary_column("df", "observation", "dog")
print(df)
}
