% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{lazyleonpackage}
\alias{lazyleonpackage}
\title{lazyleonpackage: A package for data manipulation and visualization}
\description{
The `lazyleonpackage` package provides functions for easy data manipulation and plot creation.
}
\section{Functions}{

- \code{\link{plot_variable_over_time}}: Visualizes the development of a variable over time using ggplot2.

- \code{\link{plot_correlation}}: Creates a scatter plot showing the correlation between two variables.

- \code{\link{add_binary_column}}: Adds a binary column to a data frame based on a specified value in another column.
}

