% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Scatterplot for correlation}
\usage{
plot_correlation(data, x_var, y_var)
}
\arguments{
\item{data}{A dataset as a data frame}

\item{x_var}{Name of the independent variable}

\item{y_var}{Name of the dependent variable}
}
\value{
A ggplot object with the correlation plot
}
\description{
This function creates a scatter plot to show the correlation between two variables.
}
\examples{
data <- data.frame(x = rnorm(100), y = rnorm(100))
plot_correlation(data, "x", "y")

}
