% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_variable_over_time}
\alias{plot_variable_over_time}
\title{Plot a Variable Over Time}
\usage{
plot_variable_over_time(
  data,
  date_column,
  variable_column,
  title = "Variable over Time"
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{date_column}{The name of the column representing the date or time.}

\item{variable_column}{The name of the column representing the variable to plot.}

\item{title}{The title of the plot (optional).}
}
\value{
A ggplot object representing the plot.
}
\description{
This function creates a plot to visualize the temporal evolution of a variable
from a given dataset.
}
\examples{
data <- data.frame(
  date_column = as.Date("2024-01-01") + 0:9,
  variable_column = cumsum(rnorm(10))
)
plot_variable_over_time(data, "date_column", "variable_column", "My Plot Title")
}
